+++ ariadne_graphql_proxy/contrib/aws/cache_backend.py
@@ -1,7 +1,11 @@
 import time
 from typing import Any, Optional
 
-from ariadne_graphql_proxy.cache import CacheBackend
+from ariadne_graphql_proxy.cache import (
+    CacheBackend,
+    CacheSerializer,
+    JSONCacheSerializer,
+)
 
 try:
     from asgiref.sync import sync_to_async
@@ -24,7 +28,10 @@ class DynamoDBCacheBackend(CacheBackend):
         partition_key: str = "key",
         ttl_attribute: str = "ttl",
         session: Optional[Session] = None,
+        serializer: Optional[CacheSerializer] = None,
     ) -> None:
+        super().__init__(serializer or JSONCacheSerializer())
+
         self.table_name = table_name
         self.partition_key = partition_key
         self.ttl_attribute = ttl_attribute
@@ -60,7 +67,10 @@ class DynamoDBCacheBackend(CacheBackend):
         )
 
     async def set(self, key: str, value: Any, ttl: Optional[int] = None):
-        item = {self.partition_key: key, self.value_attribute_name: value}
+        item: dict[str, Any] = {
+            self.partition_key: key,
+            self.value_attribute_name: self.serializer.serialize(value),
+        }
         if ttl:
             now = int(time.time())
             item[self.ttl_attribute] = now + ttl
@@ -74,7 +84,7 @@ class DynamoDBCacheBackend(CacheBackend):
         if len(items) < 1:
             return default
 
-        return items[0].get(self.value_attribute_name, default)
+        return self.serializer.deserialize(items[0][self.value_attribute_name])
 
     async def clear_all(self):
         pass